feat: implement shared Supabase client and update service references
