--- user_service/app/database.py
@@ -1,18 +1,0 @@
-from supabase import create_client, Client
-from app.config import settings
-import logging
-
-logger = logging.getLogger(__name__)
-
-_supabase_client: Client | None = None
-
-
-def get_supabase_client() -> Client:
-    global _supabase_client
-    if _supabase_client is None:
-        _supabase_client = create_client(settings.supabase_url, settings.supabase_key)
-    return _supabase_client
-
-
-def get_supabase() -> Client:
-    return get_supabase_client()
